Widen Term::Num from i32 to i64 to support large constants
The speech decoder graph produces dimension products exceeding i32
range (e.g. 9437184000), causing egglog expression extraction to
panic with "unsupported expression op". Widen Term::Num to i64 and
update all From impls, as_num, resolve_vars, and the egglog parser.

+++ src/shape/tracker.rs
@@ -244,7 +244,7 @@ impl ShapeTracker {
                 if let Term::Var(v) = *t
                     && let Some(val) = dyn_dim_map.get(&v)
                 {
-                    *t = Term::Num(*val as i32);
+                    *t = Term::Num(*val as i64);
                 }
             }
             d.resolve_vars(dyn_dim_map);
